+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Implement an in-memory queue for nice-only claims so the server can keep up with the increased processing speed. This takes nice-only claim endpoint times from 90-100ms to 3-5ms.
 - Add a Prometheus exporter to monitor the response times on each API endpoint
 - Fix an issue where the largest numbers (instead of nicest) are preserved during downsampling
+- Decouple detailed chunk processing size from rayon chunk processing size
 - Start using proper rust logging systems such as env_logger and test-log
 
 ## Nice v3.2.10